--- SwiftPackage-tvOS/SwiftPackage_tvOS.h
@@ -1,19 +1,0 @@
-//
-//  SwiftPackage_tvOS.h
-//  SwiftPackage-tvOS
-//
-//  Created by Khoa Pham on 25.09.2017.
-//  Copyright © 2017 <AUTHOR_NAME>. All rights reserved.
-//
-
-#import <UIKit/UIKit.h>
-
-//! Project version number for SwiftPackage_tvOS.
-FOUNDATION_EXPORT double SwiftPackage_tvOSVersionNumber;
-
-//! Project version string for SwiftPackage_tvOS.
-FOUNDATION_EXPORT const unsigned char SwiftPackage_tvOSVersionString[];
-
-// In this header, you should import all the public headers of your framework using statements like #import <SwiftPackage_tvOS/PublicHeader.h>
-
-
