--- SwiftPackage-tvOSTests/SwiftPackage_tvOSTests.swift
@@ -1,36 +1,0 @@
-//
-//  SwiftPackage_tvOSTests.swift
-//  SwiftPackage-tvOSTests
-//
-//  Created by Khoa Pham on 25.09.2017.
-//  Copyright © 2017 <AUTHOR_NAME>. All rights reserved.
-//
-
-import XCTest
-@testable import SwiftPackage_tvOS
-
-class SwiftPackage_tvOSTests: XCTestCase {
-    
-    override func setUp() {
-        super.setUp()
-        // Put setup code here. This method is called before the invocation of each test method in the class.
-    }
-    
-    override func tearDown() {
-        // Put teardown code here. This method is called after the invocation of each test method in the class.
-        super.tearDown()
-    }
-    
-    func testExample() {
-        // This is an example of a functional test case.
-        // Use XCTAssert and related functions to verify your tests produce the correct results.
-    }
-    
-    func testPerformanceExample() {
-        // This is an example of a performance test case.
-        self.measure {
-            // Put the code you want to measure the time of here.
-        }
-    }
-    
-}
