--- SwiftPackage-watchOS/SwiftPackage_watchOS.h
@@ -1,19 +1,0 @@
-//
-//  SwiftPackage_watchOS.h
-//  SwiftPackage-watchOS
-//
-//  Created by Khoa Pham on 25.09.2017.
-//  Copyright © 2017 <AUTHOR_NAME>. All rights reserved.
-//
-
-#import <WatchKit/WatchKit.h>
-
-//! Project version number for SwiftPackage_watchOS.
-FOUNDATION_EXPORT double SwiftPackage_watchOSVersionNumber;
-
-//! Project version string for SwiftPackage_watchOS.
-FOUNDATION_EXPORT const unsigned char SwiftPackage_watchOSVersionString[];
-
-// In this header, you should import all the public headers of your framework using statements like #import <SwiftPackage_watchOS/PublicHeader.h>
-
-
